pyproject.toml
[project]
name = "local-rag"
version = "0.1.0"
description = "Retrieval Augmented Generation running on a local model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.3.0",
]

[project.scripts]
generate = "main.generate:cli"
retrieve = "main.retrieve:cli"
query = "main.query:cli"